requirements.txt
llama-index
llama-index-readers-google
langchain
llama-index-vector-stores-qdrant
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai

# Apache Airflow
apache-airflow==2.8.1
apache-airflow-providers-sqlite>=3.4.3

# Additional dependencies for your ETL pipeline
pandas
python-dotenv
sqlalchemy